requirements.txt
asgiref==3.7.2
click==8.1.7
colorama==0.4.6
cssbeautifier==1.14.9
Django==4.2.4
djlint==1.32.1
django_filter==24.3    
EditorConfig==0.12.3
html-tag-names==0.1.2
html-void-elements==0.1.0
jsbeautifier==1.14.9
json5==0.9.14
pathspec==0.11.2
psycopg2-binary==2.9.7
PyYAML==6.0.1
regex==2023.8.8
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1 
tqdm==4.66.1
typing_extensions==4.7.1
django-crispy-forms==2.3
crispy-bootstrap5==2024.2
